// smartcontract/client/src/types/contracts/RabbitLaunchpad.ts
export declare namespace RabbitLaunchpad {
  export type TokenStateStruct = {
    tokenAddress: AddressLike;
    name: string;
    symbol: string;
    metadata: string;
    creator: AddressLike;
    soldSupply: BigNumberish;
    totalBNB: BigNumberish;
    initialPrice: BigNumberish;
    totalPlatformFees: BigNumberish;
    totalCreatorFees: BigNumberish;
    bondingCurveLiquidity: BigNumberish;
    liquidityPoolAmount: BigNumberish;
    graduated: boolean;
    exists: boolean;
  };

  export type TokenStateStructOutput = [
    tokenAddress: string,
    name: string,
    symbol: string,
    metadata: string,
    creator: string,
    soldSupply: bigint,
    totalBNB: bigint,
    initialPrice: bigint,
    totalPlatformFees: bigint,
    totalCreatorFees: bigint,
    bondingCurveLiquidity: bigint,
    liquidityPoolAmount: bigint,
    graduated: boolean,
    exists: boolean
  ] & {
    tokenAddress: string;
    name: string;
    symbol: string;
    metadata: string;
    creator: string;
    soldSupply: bigint;
    totalBNB: bigint;
    initialPrice: bigint;
    totalPlatformFees: bigint;
    totalCreatorFees: bigint;
    bondingCurveLiquidity: bigint;
    liquidityPoolAmount: bigint;
    graduated: boolean;
    exists: boolean;
  };
}
// ...
export interface RabbitLaunchpad extends BaseContract {
// ...
  CREATE_FEE: TypedContractMethod<[], [bigint], "view">;
// ...
  INITIAL_PRICE: TypedContractMethod<[], [bigint], "view">;
// ...
}